+++ Cursus/Libft/ft_memmove.c
@@ -11,16 +11,16 @@
 /* ************************************************************************** */
 #include"libft.h"
 
-void *ft_memmove(void *dst, const void *src, size_t len)
+void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
-	if(dst == src)
+	if (dst == src)
 		return (dst);
-	if(dst < src)
+	if (dst < src)
 	{
-		while(i < len)
+		while (i < len)
 		{
 			*((unsigned char*)dst + i) = *((unsigned char*)src + i);
 			i++;
@@ -28,7 +28,7 @@ void *ft_memmove(void *dst, const void *src, size_t len)
 	}
 	else
 	{
-		while(len > 0)
+		while (len > 0)
 		{
 			*((unsigned char*)dst + len - 1) = *((unsigned char*)src  +len - 1);
 			len--;
