+++ Cursus/Libft/ft_strncmp.c
@@ -11,12 +11,12 @@
 /* ************************************************************************** */
 #include "libft.h"
 
-int ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
-	while(i < n && s1[i] == s2[i])
+	while (i < n && s1[i] == s2[i])
 		i++;
 	return (s1[i] - s2[i]);
 }
